Python code
The ultimate goal of this project is to develop a counter-solution to account takeover attacks in payment services. The primary goal is to build a model that could identify patterns in data and spot fraudulent transactions with the highest hit-rate possible. The secondary goal is to find use of it for businesses based on their risk-aversion level:
- Send suspicious transactions to further human investigation, or
- Completely block flagged transactions